App Inventor Template
This code programs an app to translate text to speech. The app uses a text box for input, and shaking the device also gives a default message. Download this code and upload it to the MIT AppIventor website to edit. New versions should be uploaded to a different branch. Create commits and pull requests to make these edits permanent to the main branch.